Fix up packaging
We had some bad versions of deps (too tight restrictions), and
differences between gemspec in Rakefile and gemspec, which now match.
Really we should just have the gemspec, but given this is likely the
last major release of this gem, it does not seem worth the effort at
this time.

Signed-off-by: Steve Linabery <slinaber@redhat.com>