Compatibility with ActiveResource >= 3.1
This makes aeolus-image compatible with ActiveResource 3.1 and
higher while maintaining compatibility with ActiveResource 3.0.

This patch ensures that when an Aeolus::Image::Factory::* resource is
created with :id key set in hash of parameters passed into the
constructor, ActiveResource treats it like it's an existing record, not
a one to be newly created.  When calling #save on such resource
instance, ActiveResource makes a PUT request (this is how it was with
ARes 3.0 and it's the behavior expected by ImageFactory) instead of a
POST request, which would be made if the resource was treated as newly
created.